+++ src/main/java/hu/petrik/szuperhosprojekt/Szuperhos.java
@@ -2,6 +2,6 @@ package hu.petrik.szuperhosprojekt;
 
 public interface Szuperhos {
     boolean legyoziE(Szuperhos hos);
-    int mekkoraAzEreje();
+    double mekkoraAzEreje();
     void gyogyul();
 }
